rename and move two methods
