+++ sql/catalyst/src/test/scala/org/apache/spark/sql/catalyst/optimizer/joinReorder/JoinReorderSuite.scala
@@ -130,7 +130,7 @@ class JoinReorderSuite extends JoinReorderPlanTestBase with StatsEstimationTestB
         .join(t2, Inner, Some(nameToAttr("t1.k-1-2") === nameToAttr("t2.k-1-5")))
         .select(outputsOf(t1, t2, t3): _*)
 
-    assertEqualPlans(Optimize, originalPlan, bestPlan)
+    assertEqualJoinPlans(Optimize, originalPlan, bestPlan)
   }
 
   test("put unjoinable item at the end and reorder 3 joinable tables") {
@@ -147,7 +147,7 @@ class JoinReorderSuite extends JoinReorderPlanTestBase with StatsEstimationTestB
         .join(t4)
         .select(outputsOf(t1, t2, t4, t3): _*)
 
-    assertEqualPlans(Optimize, originalPlan, bestPlan)
+    assertEqualJoinPlans(Optimize, originalPlan, bestPlan)
   }
 
   test("reorder 3 tables with pure-attribute project") {
@@ -162,7 +162,7 @@ class JoinReorderSuite extends JoinReorderPlanTestBase with StatsEstimationTestB
         .join(t2, Inner, Some(nameToAttr("t1.k-1-2") === nameToAttr("t2.k-1-5")))
         .select(nameToAttr("t1.v-1-10"))
 
-    assertEqualPlans(Optimize, originalPlan, bestPlan)
+    assertEqualJoinPlans(Optimize, originalPlan, bestPlan)
   }
 
   test("reorder 3 tables - one of the leaf items is a project") {
@@ -179,7 +179,7 @@ class JoinReorderSuite extends JoinReorderPlanTestBase with StatsEstimationTestB
           Some(nameToAttr("t1.k-1-2") === nameToAttr("t5.k-1-5")))
         .select(nameToAttr("t1.v-1-10"))
 
-    assertEqualPlans(Optimize, originalPlan, bestPlan)
+    assertEqualJoinPlans(Optimize, originalPlan, bestPlan)
   }
 
   test("don't reorder if project contains non-attribute") {
@@ -189,7 +189,7 @@ class JoinReorderSuite extends JoinReorderPlanTestBase with StatsEstimationTestB
         .join(t3, Inner, Some(nameToAttr("t1.v-1-10") === nameToAttr("t3.v-1-100")))
         .select("key".attr)
 
-    assertEqualPlans(Optimize, originalPlan, originalPlan)
+    assertEqualJoinPlans(Optimize, originalPlan, originalPlan)
   }
 
   test("reorder 4 tables (bushy tree)") {
@@ -210,7 +210,7 @@ class JoinReorderSuite extends JoinReorderPlanTestBase with StatsEstimationTestB
           Inner, Some(nameToAttr("t1.k-1-2") === nameToAttr("t4.k-1-2")))
         .select(outputsOf(t1, t4, t2, t3): _*)
 
-    assertEqualPlans(Optimize, originalPlan, bestPlan)
+    assertEqualJoinPlans(Optimize, originalPlan, bestPlan)
   }
 
   test("keep the order of attributes in the final output") {
@@ -222,7 +222,7 @@ class JoinReorderSuite extends JoinReorderPlanTestBase with StatsEstimationTestB
           .join(t2, Inner, Some(nameToAttr("t1.k-1-2") === nameToAttr("t2.k-1-5")))
           .select(expectedOrder: _*)
       // The plan should not change after optimization
-      assertEqualPlans(Optimize, expectedPlan, expectedPlan)
+      assertEqualJoinPlans(Optimize, expectedPlan, expectedPlan)
     }
   }
 
@@ -240,7 +240,7 @@ class JoinReorderSuite extends JoinReorderPlanTestBase with StatsEstimationTestB
           .join(tab2, Cross, Some('b === 'i))
           .select(outputsOf(tab1, tab2, tab3): _*)
 
-    assertEqualPlans(Optimize, original, expected)
+    assertEqualJoinPlans(Optimize, original, expected)
   }
 
   test("reorder recursively") {
@@ -284,7 +284,7 @@ class JoinReorderSuite extends JoinReorderPlanTestBase with StatsEstimationTestB
       .union(t4.select(nameToAttr("t4.v-1-10")))
       .join(t5, Inner, Some(nameToAttr("t1.v-1-10") === nameToAttr("t5.v-1-5")))
 
-    assertEqualPlans(Optimize, originalPlan, bestPlan)
+    assertEqualJoinPlans(Optimize, originalPlan, bestPlan)
   }
 
   test("don't reorder if hints present") {
@@ -297,7 +297,7 @@ class JoinReorderSuite extends JoinReorderPlanTestBase with StatsEstimationTestB
           Inner,
           Some(nameToAttr("t1.k-1-2") === nameToAttr("t4.k-1-2")))
 
-    assertEqualPlans(Optimize, originalPlan, originalPlan)
+    assertEqualJoinPlans(Optimize, originalPlan, originalPlan)
 
     val originalPlan2 =
       t1.join(t2, Inner, Some(nameToAttr("t1.k-1-2") === nameToAttr("t2.k-1-5")))
@@ -306,7 +306,7 @@ class JoinReorderSuite extends JoinReorderPlanTestBase with StatsEstimationTestB
         .hint("broadcast")
         .join(t3, Inner, Some(nameToAttr("t4.v-1-10") === nameToAttr("t3.v-1-100")))
 
-    assertEqualPlans(Optimize, originalPlan2, originalPlan2)
+    assertEqualJoinPlans(Optimize, originalPlan2, originalPlan2)
 
     val originalPlan3 =
       t1.join(t2, Inner, Some(nameToAttr("t1.k-1-2") === nameToAttr("t2.k-1-5")))
@@ -314,7 +314,7 @@ class JoinReorderSuite extends JoinReorderPlanTestBase with StatsEstimationTestB
         .join(t3, Inner, Some(nameToAttr("t1.v-1-10") === nameToAttr("t3.v-1-100")))
         .join(t5, Inner, Some(nameToAttr("t5.v-1-5") === nameToAttr("t3.v-1-100")))
 
-    assertEqualPlans(Optimize, originalPlan3, originalPlan3)
+    assertEqualJoinPlans(Optimize, originalPlan3, originalPlan3)
   }
 
   test("reorder below and above the hint node") {
@@ -330,7 +330,7 @@ class JoinReorderSuite extends JoinReorderPlanTestBase with StatsEstimationTestB
       .select(outputsOf(t1, t2, t3): _*)
       .hint("broadcast").join(t4)
 
-    assertEqualPlans(Optimize, originalPlan, bestPlan)
+    assertEqualJoinPlans(Optimize, originalPlan, bestPlan)
 
     val originalPlan2 =
       t1.join(t2).join(t3)
@@ -344,7 +344,7 @@ class JoinReorderSuite extends JoinReorderPlanTestBase with StatsEstimationTestB
         .select(outputsOf(t1, t2, t3): _*)
         .join(t4.hint("broadcast"))
 
-    assertEqualPlans(Optimize, originalPlan2, bestPlan2)
+    assertEqualJoinPlans(Optimize, originalPlan2, bestPlan2)
 
     val originalPlan3 =
       t1.join(t2, Inner, Some(nameToAttr("t1.k-1-2") === nameToAttr("t2.k-1-5")))
@@ -361,6 +361,6 @@ class JoinReorderSuite extends JoinReorderPlanTestBase with StatsEstimationTestB
         .join(t4, Inner, Some(nameToAttr("t4.v-1-10") === nameToAttr("t3.v-1-100")))
         .join(t5, Inner, Some(nameToAttr("t5.v-1-5") === nameToAttr("t3.v-1-100")))
 
-    assertEqualPlans(Optimize, originalPlan3, bestPlan3)
+    assertEqualJoinPlans(Optimize, originalPlan3, bestPlan3)
   }
 }
